app/requirements.txt
Bootstrap-Flask
Flask-Script==2.0.5
flask-appbuilder
Flask-SQLAlchemy==2.3
Flask-WTF==0.14
gunicorn==19.4.5
itsdangerous==0.24
pytz==2016.10
structlog==16.1.0
termcolor==1.1.0
WTForms==2.2.1
Flask==1.0.2
Flask-Admin==1.3.0
six
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.0